refactor: move MessageSendFailureHandler.kt to `domain` package
