+++ app/src/test/kotlin/com/wire/android/feature/conversation/content/domain/MessageSendFailureHandlerTest.kt
@@ -1,4 +1,4 @@
-package com.wire.android.feature.conversation.content.usecase
+package com.wire.android.feature.conversation.content.domain
 
 import com.wire.android.UnitTest
 import com.wire.android.core.exception.Failure
